+++ CHANGES.rst
@@ -5,6 +5,7 @@
 - Add column exists check to bioutils.pseudobulk_table (#356)
 - Fix main title in receptor-ligand network plot (#358)
 - load_h5ad(): warn if adata.raw is found. (#352)
+- receptor-ligand: nan zscore to 0 (#302)
 
 Changes to notebooks
 ^^^^^^^^^^^^^^^^^^^^